fix: 修复了dpkg日志和xorg日志加载数据时阻塞界面的问题和调整了导出日志文件默认目录为桌面,并且消除了代码中的所有warning
Description: 把xorg日志和dpkg日志读取文件操作放入子线程中，现在xorg日志，kwin日志，dpkg日志，内核日志和启动日志均在同一线程实例中加载数据

Log: fix: 修复了dpkg日志和xorg日志加载数据时阻塞界面的问题和调整了导出日志文件默认目录为桌面,并且消除了代码中的所有warning

link: https://pms.uniontech.com/zentao/bug-view-32344.html
      https://pms.uniontech.com/zentao/bug-view-32638.html

+++ 3rdparty/docx/length.cpp
@@ -25,7 +25,7 @@ Length::Length(int emu) : m_value(emu)
 
 float Length::cm() const
 {
-    return (float)m_value/(float)EMUS_PER_CM;
+    return static_cast<float>(m_value) / static_cast<float>(EMUS_PER_CM);
 }
 
 int Length::emu() const
@@ -35,22 +35,22 @@ int Length::emu() const
 
 float Length::inches() const
 {
-    return (float)m_value/(float)EMUS_PER_INCH;
+    return static_cast<float>(m_value) / static_cast<float>(EMUS_PER_INCH);
 }
 
 float Length::mm() const
 {
-    return (float)m_value/(float)EMUS_PER_MM;
+    return static_cast<float>(m_value) / static_cast<float>(EMUS_PER_MM);
 }
 
 int Length::px() const
 {
-    return (int)(round((float)m_value/(float)EMUS_PER_PX) + 0.1);
+    return  static_cast<int>(round(static_cast<float>(m_value) / static_cast<float>(EMUS_PER_PX)) + static_cast<float>(0.1));
 }
 
 int Length::twips() const
 {
-    return (int)(round((float)m_value/(float)EMUS_PER_TWIP));
+    return static_cast<int>(round(static_cast<float>(m_value) / static_cast<float>(EMUS_PER_TWIP)));
 }
 
 bool Length::isEmpty() const
@@ -93,7 +93,7 @@ Length Cm::emus(float cm)
  */
 Length Emu::emus(float emu)
 {
-    return Length(emu);
+    return Length(static_cast<int>(emu));
 }
 
 /*!
